added KanbanBoard.tsx component
